requirements.txt
asgiref==3.6.0
Django==4.2.1
django-bootstrap5==23.1
djangorestframework==3.14.0
mysqlclient==2.1.1
sqlparse==0.4.4
tzdata==2023.3
